bin/public-inbox-tg: Skip the test robot content and ignore the Android tree
Don't show the content of the email if it comes from the kernel test
robot. Also, don't send a notification if the error is related to the
Android tree.

Android tree is full of errors when it's tested by the Intel's kernel
test robot. One of the reason is that, Android tree only cares to be
compiled with specific configuration and archs, they don't fully cover
all the mainline testing.

Signed-off-by: Ammar Faizi <ammarfaizi2@gnuweeb.org>