bin/public-inbox-tg: Fix `extractList()` function
When the "To" or "CC" list contains an address that has "," character in
the name, the resulting array list is wrong because this function assumes
"," character as an address separator. For example, the email address
`"Chen, Rong A" <rong.a.chen@intel.com>` is split into two addresses due
to the comma character that sits after the "Chen".

Fix this by enumerating the list with a loop and a state variable, parse
them manually.

Fixes: 962100a0051db717f4053ffe91a41e3cee3434be ("Initial commit for public inbox")
Co-authored-by: Nugra <richiisei@gmail.com>
Signed-off-by: Nugra <richiisei@gmail.com>
Signed-off-by: Alviro Iskandar Setiawan <alviro.iskandar@gnuweeb.org>
Link: https://lore.gnuweeb.org/gwml/20220305173617.3605062-4-alviro.iskandar@gnuweeb.org
Signed-off-by: Ammar Faizi <ammarfaizi2@gnuweeb.org>