bin/public-inbox-tg: Fix regex pattern for headers
When testing email with Faris, we found that Faris' emails don't get
processed. They appear on the lore, but don't appear on Telegram. Sir
Ammar debugged this bug and the log said: "Cannot get the \"message-id\"
line".

It turned out that the regex pattern for headers was wrong, if the
header to be matched happens to be at the end of the headers, it will
not have a double LF because previously we have already split the
double LF with explode(). Fix this by changing the pattern to match the
header if it happens to be at the end of the headers.

Cc: Faris Suryanto <onlyfaris@gnuweeb.org>
Debugged-by: Ammar Faizi <ammarfaizi2@gnuweeb.org>
Signed-off-by: Alviro Iskandar Setiawan <alviro.iskandar@gnuweeb.org>
Link: https://lore.gnuweeb.org/gwml/20220306130024.520329-1-alviro.iskandar@gnuweeb.org
Signed-off-by: Ammar Faizi <ammarfaizi2@gnuweeb.org>